+++ library/platform.c
@@ -53,7 +53,7 @@ static void platform_free_uninit( void *ptr )
 void * (*mbedtls_calloc)( size_t, size_t ) = MBEDTLS_PLATFORM_STD_CALLOC;
 void (*mbedtls_free)( void * )     = MBEDTLS_PLATFORM_STD_FREE;
 
-int mbedtls_platform_set_calloc_free( void * (*calloc_func)( size_t, size_t ),
+int32_t mbedtls_platform_set_calloc_free( void * (*calloc_func)( size_t, size_t ),
                               void (*free_func)( void * ) )
 {
     mbedtls_calloc = calloc_func;
@@ -64,9 +64,9 @@ int mbedtls_platform_set_calloc_free( void * (*calloc_func)( size_t, size_t ),
 
 #if defined(_WIN32)
 #include <stdarg.h>
-int mbedtls_platform_win32_snprintf( char *s, size_t n, const char *fmt, ... )
+int32_t mbedtls_platform_win32_snprintf( char *s, size_t n, const char *fmt, ... )
 {
-    int ret;
+    int32_t ret;
     va_list argp;
 
     /* Avoid calling the invalid parameter handler by checking ourselves */
@@ -95,7 +95,7 @@ int mbedtls_platform_win32_snprintf( char *s, size_t n, const char *fmt, ... )
 /*
  * Make dummy function to prevent NULL pointer dereferences
  */
-static int platform_snprintf_uninit( char * s, size_t n,
+static int32_t platform_snprintf_uninit( char * s, size_t n,
                                      const char * format, ... )
 {
     ((void) s);
@@ -107,11 +107,11 @@ static int platform_snprintf_uninit( char * s, size_t n,
 #define MBEDTLS_PLATFORM_STD_SNPRINTF    platform_snprintf_uninit
 #endif /* !MBEDTLS_PLATFORM_STD_SNPRINTF */
 
-int (*mbedtls_snprintf)( char * s, size_t n,
+int32_t (*mbedtls_snprintf)( char * s, size_t n,
                           const char * format,
                           ... ) = MBEDTLS_PLATFORM_STD_SNPRINTF;
 
-int mbedtls_platform_set_snprintf( int (*snprintf_func)( char * s, size_t n,
+int32_t mbedtls_platform_set_snprintf( int32_t (*snprintf_func)( char * s, size_t n,
                                                  const char * format,
                                                  ... ) )
 {
@@ -125,7 +125,7 @@ int mbedtls_platform_set_snprintf( int (*snprintf_func)( char * s, size_t n,
 /*
  * Make dummy function to prevent NULL pointer dereferences
  */
-static int platform_printf_uninit( const char *format, ... )
+static int32_t platform_printf_uninit( const char *format, ... )
 {
     ((void) format);
     return( 0 );
@@ -134,9 +134,9 @@ static int platform_printf_uninit( const char *format, ... )
 #define MBEDTLS_PLATFORM_STD_PRINTF    platform_printf_uninit
 #endif /* !MBEDTLS_PLATFORM_STD_PRINTF */
 
-int (*mbedtls_printf)( const char *, ... ) = MBEDTLS_PLATFORM_STD_PRINTF;
+int32_t (*mbedtls_printf)( const char *, ... ) = MBEDTLS_PLATFORM_STD_PRINTF;
 
-int mbedtls_platform_set_printf( int (*printf_func)( const char *, ... ) )
+int32_t mbedtls_platform_set_printf( int32_t (*printf_func)( const char *, ... ) )
 {
     mbedtls_printf = printf_func;
     return( 0 );
@@ -148,7 +148,7 @@ int mbedtls_platform_set_printf( int (*printf_func)( const char *, ... ) )
 /*
  * Make dummy function to prevent NULL pointer dereferences
  */
-static int platform_fprintf_uninit( FILE *stream, const char *format, ... )
+static int32_t platform_fprintf_uninit( FILE *stream, const char *format, ... )
 {
     ((void) stream);
     ((void) format);
@@ -158,10 +158,10 @@ static int platform_fprintf_uninit( FILE *stream, const char *format, ... )
 #define MBEDTLS_PLATFORM_STD_FPRINTF   platform_fprintf_uninit
 #endif /* !MBEDTLS_PLATFORM_STD_FPRINTF */
 
-int (*mbedtls_fprintf)( FILE *, const char *, ... ) =
+int32_t (*mbedtls_fprintf)( FILE *, const char *, ... ) =
                                         MBEDTLS_PLATFORM_STD_FPRINTF;
 
-int mbedtls_platform_set_fprintf( int (*fprintf_func)( FILE *, const char *, ... ) )
+int32_t mbedtls_platform_set_fprintf( int32_t (*fprintf_func)( FILE *, const char *, ... ) )
 {
     mbedtls_fprintf = fprintf_func;
     return( 0 );
@@ -173,7 +173,7 @@ int mbedtls_platform_set_fprintf( int (*fprintf_func)( FILE *, const char *, ...
 /*
  * Make dummy function to prevent NULL pointer dereferences
  */
-static void platform_exit_uninit( int status )
+static void platform_exit_uninit( int32_t status )
 {
     ((void) status);
 }
@@ -181,9 +181,9 @@ static void platform_exit_uninit( int status )
 #define MBEDTLS_PLATFORM_STD_EXIT   platform_exit_uninit
 #endif /* !MBEDTLS_PLATFORM_STD_EXIT */
 
-void (*mbedtls_exit)( int status ) = MBEDTLS_PLATFORM_STD_EXIT;
+void (*mbedtls_exit)( int32_t status ) = MBEDTLS_PLATFORM_STD_EXIT;
 
-int mbedtls_platform_set_exit( void (*exit_func)( int status ) )
+int32_t mbedtls_platform_set_exit( void (*exit_func)( int32_t status ) )
 {
     mbedtls_exit = exit_func;
     return( 0 );
@@ -208,7 +208,7 @@ static mbedtls_time_t platform_time_uninit( mbedtls_time_t* timer )
 
 mbedtls_time_t (*mbedtls_time)( mbedtls_time_t* timer ) = MBEDTLS_PLATFORM_STD_TIME;
 
-int mbedtls_platform_set_time( mbedtls_time_t (*time_func)( mbedtls_time_t* timer ) )
+int32_t mbedtls_platform_set_time( mbedtls_time_t (*time_func)( mbedtls_time_t* timer ) )
 {
     mbedtls_time = time_func;
     return( 0 );
@@ -222,7 +222,7 @@ int mbedtls_platform_set_time( mbedtls_time_t (*time_func)( mbedtls_time_t* time
 /* Default implementations for the platform independent seed functions use
  * standard libc file functions to read from and write to a pre-defined filename
  */
-int mbedtls_platform_std_nv_seed_read( unsigned char *buf, size_t buf_len )
+int32_t mbedtls_platform_std_nv_seed_read( unsigned char *buf, size_t buf_len )
 {
     FILE *file;
     size_t n;
@@ -237,10 +237,10 @@ int mbedtls_platform_std_nv_seed_read( unsigned char *buf, size_t buf_len )
     }
 
     fclose( file );
-    return( (int)n );
+    return( n );
 }
 
-int mbedtls_platform_std_nv_seed_write( unsigned char *buf, size_t buf_len )
+int32_t mbedtls_platform_std_nv_seed_write( unsigned char *buf, size_t buf_len )
 {
     FILE *file;
     size_t n;
@@ -255,7 +255,7 @@ int mbedtls_platform_std_nv_seed_write( unsigned char *buf, size_t buf_len )
     }
 
     fclose( file );
-    return( (int)n );
+    return( n );
 }
 #endif /* MBEDTLS_PLATFORM_NO_STD_FUNCTIONS */
 
@@ -264,7 +264,7 @@ int mbedtls_platform_std_nv_seed_write( unsigned char *buf, size_t buf_len )
 /*
  * Make dummy function to prevent NULL pointer dereferences
  */
-static int platform_nv_seed_read_uninit( unsigned char *buf, size_t buf_len )
+static int32_t platform_nv_seed_read_uninit( unsigned char *buf, size_t buf_len )
 {
     ((void) buf);
     ((void) buf_len);
@@ -278,7 +278,7 @@ static int platform_nv_seed_read_uninit( unsigned char *buf, size_t buf_len )
 /*
  * Make dummy function to prevent NULL pointer dereferences
  */
-static int platform_nv_seed_write_uninit( unsigned char *buf, size_t buf_len )
+static int32_t platform_nv_seed_write_uninit( unsigned char *buf, size_t buf_len )
 {
     ((void) buf);
     ((void) buf_len);
@@ -288,14 +288,14 @@ static int platform_nv_seed_write_uninit( unsigned char *buf, size_t buf_len )
 #define MBEDTLS_PLATFORM_STD_NV_SEED_WRITE   platform_nv_seed_write_uninit
 #endif /* !MBEDTLS_PLATFORM_STD_NV_SEED_WRITE */
 
-int (*mbedtls_nv_seed_read)( unsigned char *buf, size_t buf_len ) =
+int32_t (*mbedtls_nv_seed_read)( unsigned char *buf, size_t buf_len ) =
             MBEDTLS_PLATFORM_STD_NV_SEED_READ;
-int (*mbedtls_nv_seed_write)( unsigned char *buf, size_t buf_len ) =
+int32_t (*mbedtls_nv_seed_write)( unsigned char *buf, size_t buf_len ) =
             MBEDTLS_PLATFORM_STD_NV_SEED_WRITE;
 
-int mbedtls_platform_set_nv_seed(
-        int (*nv_seed_read_func)( unsigned char *buf, size_t buf_len ),
-        int (*nv_seed_write_func)( unsigned char *buf, size_t buf_len ) )
+int32_t mbedtls_platform_set_nv_seed(
+        int32_t (*nv_seed_read_func)( unsigned char *buf, size_t buf_len ),
+        int32_t (*nv_seed_write_func)( unsigned char *buf, size_t buf_len ) )
 {
     mbedtls_nv_seed_read = nv_seed_read_func;
     mbedtls_nv_seed_write = nv_seed_write_func;
